move out api/rpc
