+++ src/common/metrics/src/lib.rs
@@ -51,6 +51,7 @@ pub use crate::metrics::cluster;
 /// Metrics.
 pub use crate::metrics::http;
 pub use crate::metrics::interpreter;
+pub use crate::metrics::mysql;
 pub use crate::metrics::openai;
 pub use crate::metrics::session;
 pub use crate::metrics::storage;
