+++ src/query/service/Cargo.toml
@@ -28,7 +28,6 @@ io-uring = [
     # "common-meta-raft-store/io-uring",
 ]
 
-
 [dependencies]
 # Workspace dependencies
 common-arrow = { path = "../../common/arrow" }
