+++ src/query/service/src/api/rpc/exchange/exchange_source_reader.rs
@@ -19,6 +19,7 @@ use std::sync::Arc;
 
 use common_exception::Result;
 use common_expression::DataBlock;
+use common_metrics::transform::*;
 use common_pipeline_core::pipe::Pipe;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
@@ -31,11 +32,9 @@ use common_pipeline_core::Pipeline;
 use common_pipeline_transforms::processors::transforms::TransformDummy;
 use log::info;
 
-use crate::api::rpc::exchange::metrics::metrics_inc_exchange_read_bytes;
-use crate::api::rpc::exchange::metrics::metrics_inc_exchange_read_count;
-use crate::api::rpc::exchange::serde::exchange_deserializer::ExchangeDeserializeMeta;
 use crate::api::rpc::flight_client::FlightReceiver;
 use crate::api::DataPacket;
+use crate::api::ExchangeDeserializeMeta;
 
 pub struct ExchangeSourceReader {
     finished: AtomicBool,
