+++ src/query/service/src/api/rpc/exchange/mod.rs
@@ -24,7 +24,6 @@ mod exchange_source_reader;
 mod exchange_transform;
 mod exchange_transform_scatter;
 mod exchange_transform_shuffle;
-mod metrics;
 mod serde;
 mod statistics_receiver;
 mod statistics_sender;
