+++ src/query/service/src/servers/mysql/mod.rs
@@ -15,7 +15,6 @@
 mod mysql_federated;
 mod mysql_handler;
 mod mysql_interactive_worker;
-mod mysql_metrics;
 mod mysql_session;
 #[allow(clippy::unused_io_amount)]
 mod reject_connection;
